requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
geopy==2.4.1
requests==2.32.3